Let's implement the ability to send the emails from the server.
First we need to get the right email address for the reply-to and
can't assume users@ like we did before; so parse the list from ponymail
but only do it when we need it (cache it if the doc is saved)

Change the way we render the email; rather than doing it in HTML and
parsting to text, just build it all up in text.

Provide a function on the server that will send the emails; since all
our users are trusted we don't need to verify they didn't mess with the
fields that are submitted as hidden text boxes; the mails will be sent
out as them.  The CSRF token should help block any cunning attacks.

Switch to a single tab for both/all the emails too, much less confusing.