Validation for severity level
This adds custom validation that when an 'other' metrics is added,
it needs to be filled.

Shouldn't this already have been caught by the 'default' validation
via the schema though?

This does not yet validate that there is exactly one metric, that
that metric is of type 'other', or that the value is one of the four
common severity levels.