send emails after saving the CVE
Previously, when using the 'Send these Emails' button on the
'OSS/ASF Emails' tab, the CVE was saved with the 'CNA_private.emailed'
flag enabled, and in parallel the form to trigger the emails was
submitted.

This change makes those two actions sequential: first save the CVE
(with the 'emailed' flag set to 'yes'), then post the form.

This fixes https://github.com/apache/security-vulnogram/issues/115 ,
both the fact that an error box was briefly shown, and
the fact that sometimes the 'emailed' flag would not get set to 'yes'
correctly.

A downside of this issue is that in case of failure to send the
emails, the 'emailed' flag will still be set to 'yes'. That seems
acceptable to me: I don't think it has happened much, and when it
does it's easy enough to reset the flag on the 'Source' tab.