problemType: autocomplete all CWEs
The NVD has clarified that they highly encourage us to begin
providing CWE values that are outside the CWE-1003 to be more
granular with our CWE assertions. Should there be clarity concerns
or improper negative impacts to our CVMAP assessment results, we
can discuss those as they arise and make adjustments to our
understandings or processes as needed.